GCP/1.4 PubSub/requirements.txt
flask
google-cloud-firestore
google-cloud-pubsub
numpy
requests
gunicorn
